- Refactored headers and code to allow SDK to build using new _WIN32, PLATFORM_WINDOWS_EMULATOR split
